Add test cases for runner container.
  Add: A simple runner for testing purposes.
  Add: Test cases for runner container

+++ st2tests/st2tests/base.py
@@ -55,7 +55,8 @@ class DbTestCase(TestCase):
     @classmethod
     def tearDownClass(cls):
         cls.drop_collections()
-        DbTestCase.db_connection.drop_database(cfg.CONF.database.db_name)
+        if DbTestCase.db_connection is not None:
+            DbTestCase.db_connection.drop_database(cfg.CONF.database.db_name)
         db_teardown()
         DbTestCase.db_connection = None
